feat: add custom CSS for Chzzk chat in 4K UHD
This is my custom CSS file for a Chzzk chat window added as a browser
source to OBS where the base (canvas) resolution is 3840 × 2160 (4K
UHD).

The file uses two fonts: Victor Mono and Noto Sans CJK KR. Feel free to
replace them with the fonts you prefer.

I added a black background with an opacity of 0.9 to each item in the
chat list; the background has rounded corners with a radius of 24
pixels.

I enlarged all elements in the chat, subscription, and donation messages
by a scale factor of 4.

Co-authored-by: RessChild <39401832+RessChild@users.noreply.github.com>